+++ src/watcher.rs
@@ -29,7 +29,7 @@ pub struct Watcher {
 impl Watcher {
     pub fn new(
         paths_to_watch: &[PathBuf],
-        mut ignorer: Option<Ignorer>,
+        mut ignorer: IgnorerSet,
     ) -> Result<Self> {
         let (sender, receiver) = bounded(0);
         let mut notify_watcher =
@@ -55,18 +55,16 @@ impl Watcher {
                             debug!("notify event: {we:?}");
                         }
                     }
-                    if let Some(ignorer) = ignorer.as_mut() {
-                        match time!(Info, ignorer.excludes_all(&we.paths)) {
-                            Ok(true) => {
-                                debug!("all excluded");
-                                return;
-                            }
-                            Ok(false) => {
-                                debug!("at least one is included");
-                            }
-                            Err(e) => {
-                                warn!("exclusion check failed: {e}");
-                            }
+                    match time!(Info, ignorer.excludes_all_pathbufs(&we.paths)) {
+                        Ok(true) => {
+                            debug!("all excluded");
+                            return;
+                        }
+                        Ok(false) => {
+                            debug!("at least one is included");
+                        }
+                        Err(e) => {
+                            warn!("exclusion check failed: {e}");
                         }
                     }
                     if let Err(e) = sender.send(()) {