+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix Miri output seen as wrong when there's only warnings
 - more lenient detection of warnings and errors due to 'miri run' not supporting `--color` - Fix #251
 - eslint analyzer (set `analyzer = "eslint"` in your job definition)
+- new `ignore` job parameter, accepts a list of glob patterns
 
 <a name="v3.2.0"></a>
 ### v3.2.0 - 2024/11/04